House: Chaining the models City Country
I removed the countries selection logic from the HTML file, and replace it by the lib "django-smart-select"

this lib chains models with jquery fetching automatically in the admin and in any form as seen in the docs.

the latest version of this lib(1.6.0) is the first supporting python 3.9 which we are using but this version isn't released yet. so the solution i saw people use is to explicitly state that we are using the lib from the version of the repo in github see Pipfile.

and thus no more worrying about logic in the HTML.

Because all this magic is happening from the 3rd party library there is no real code to test that isn't tested thoroughly in the library itself, so I don't think any new tests need to be created If there is I would like to hear the idea of the test because the output of this PR is still "frontendy" meaning changes are happening to the HTML element "select" and its options so I cant think of a test that tests the code behind that. But I checked this on the House filter form and in the admin and this works perfectly

The tests changed only because we're creating the initial countries and cities data using the django command loaddata, so the tests need to run that first as a fixture. and that created some problems with the test classes so I moved the problematic ones to be functional and that fixed the problem.